Drop Rails 5.2 support
