+++ ransack.gemspec
@@ -15,8 +15,8 @@ Gem::Specification.new do |s|
   s.required_ruby_version = '>= 2.6'
   s.license     = 'MIT'
 
-  s.add_dependency 'activerecord', '>= 5.2.4'
-  s.add_dependency 'activesupport', '>= 5.2.4'
+  s.add_dependency 'activerecord', '>= 6.0.4'
+  s.add_dependency 'activesupport', '>= 6.0.4'
   s.add_dependency 'i18n'
 
   s.files         = `git ls-files`.split("\n")
